shex_ast::ast::serde_string_or_struct
Function deserialize_string_or_struct

pub fn deserialize_string_or_struct<'de, T, D>(d: D) -> Result<T, D::Error>
where
    T: Deserialize<'de> + FromStr,
    <T as FromStr>::Err: Display,
    D: Deserializer<'de>,

Handle a value which may either a struct that deserializes as type T, or a bare string that can be turned into a type T using FromStr::from_str. We do this in a clever way that allows us to be generic over multiple such types, and which allows us to use the structure deserialization code automatically generated by serde.

An earlier and much uglier version of this parser was the inspiration for this official serde example, which in turn forms the basis of this code.