shex_ast::ir::schema_ir

Struct SchemaIR

pub struct SchemaIR { /* private fields */ }

Implementations

impl SchemaIR

pub fn new() -> SchemaIR

pub fn set_prefixmap(&mut self, prefixmap: Option<PrefixMap>)

pub fn prefixmap(&self) -> PrefixMap

pub fn add_shape(&mut self, shape_label: ShapeLabel, se: ShapeExpr)

pub fn get_shape_expr(&self, shape_label: &ShapeLabel) -> Option<&ShapeExpr>

pub fn from_schema_json( &mut self, schema_json: &SchemaJson, ) -> Result<(), SchemaIRError>

pub fn find_ref( &self, se_ref: &ShapeExprLabel, ) -> Result<ShapeLabelIdx, SchemaIRError>

pub fn find_label( &self, label: &ShapeLabel, ) -> Option<(&ShapeLabelIdx, &ShapeExpr)>

pub fn find_shape_label_idx(&self, label: &ShapeLabel) -> Option<&ShapeLabelIdx>

pub fn find_shape_idx( &self, idx: &ShapeLabelIdx, ) -> Option<&(Option<ShapeLabel>, ShapeExpr)>

pub fn shape_label_from_idx(&self, idx: &ShapeLabelIdx) -> Option<&ShapeLabel>

pub fn new_index(&mut self) -> ShapeLabelIdx

pub fn existing_labels(&self) -> Vec<&ShapeLabel>

pub fn shapes(&self) -> impl Iterator<Item = &(Option<ShapeLabel>, ShapeExpr)>

pub fn references( &self, idx: &ShapeLabelIdx, ) -> HashMap<Pred, Vec<ShapeLabelIdx>>

pub fn references_visited( &self, idx: &ShapeLabelIdx, visited: HashSet<ShapeLabelIdx>, ) -> HashMap<Pred, Vec<ShapeLabelIdx>>

pub fn get_shape_label_idx( &self, shape_label: &ShapeLabel, ) -> Result<ShapeLabelIdx, SchemaIRError>

pub fn replace_shape(&mut self, idx: &ShapeLabelIdx, se: ShapeExpr)

pub fn show_label(&self, label: &ShapeLabel) -> String

pub fn neg_cycles( &self, ) -> Vec<Vec<(ShapeLabelIdx, ShapeLabelIdx, Vec<ShapeLabelIdx>)>>

pub fn has_neg_cycle(&self) -> bool

This is used to detect cycles that involve negations in the schema A well formed schema should not have any cyclic reference that involve a negation

pub fn dependencies(&self) -> Vec<(ShapeLabel, PosNeg, ShapeLabel)>